requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
openai==1.57.2
groq==0.11.0
transformers==4.47.1
torch==2.5.1
numpy==1.26.4
pydantic==2.10.3

